Feat: Add support for submodules
Why do we need this change?
=======================
Submodules are a function that teams leverage for inclusion and updating in various ways. As submodules comes from the .gitmodules file and has a particular format we need to parse it and gather the path for each one to generate a configuration for it.

What effects does this change have?
=======================
* Adds new dependency for gitcommit rust parser
* updates found_to_update to support the .gitmodule structure in addition to the standard behaviors for others